requirements.txt
xgboost
pandas
numpy
scikit-learn
pyserial
matplotlib
joblib
